Refactor import paths for habit-related components
