+++ app/habits/[habitId]/edit/index.tsx
@@ -12,7 +12,7 @@ import { Crosshair } from "~/lib/icons/Crosshair";
 import { Sun } from "~/lib/icons/Sun";
 import { Bell } from "~/lib/icons/Bell";
 import { ChevronRight } from "~/lib/icons/ChevronRight";
-import ScheduleStartDate from "~/components/schedule-start-date";
+import ScheduleStartDate from "~/components/habits/schedule-start-date";
 import {
   type DailyRepeat,
   type Recurrence,
